Fix caching of builtins, prepare caching for has_avatar.
There were a few problems with this library. It's possible that they're
resolved in a more current version of Django or Python, but in the
versions we're stuck with, caching is completely broken.

First, the decorators were being applied in the wrong order. Each
template tag has two decorators - one that transforms the function
into a cached version, and one that creates the link to the
template tag. The problem was that the link to the template tag
was added /before/ the caching was enabled. Reversing the order
fixed that (making the memoized version the tag), but introduced...

An issue where the "register.simple_tag" apparently needs you to
specify the name if it's not the innermost decorator. This was
weird, but easy to fix by adding `name=` tags.

Third, I've added a `_0arg` version of the caching decorator, and
applied it to `has_avatar`. This doesn't actually impact JudgeApps,
because we're using a version that is inside our own repo, but
creating the decorator means that I can make a later patch to
JudgeApps that will also enable caching on that function.

Fourth, I noticed that the list of "cached functions" inside
django-avatar is populated when the function is called, not when
it is decorated. Since this list is used to invalidate the cache
later, it's important to build this list as django-avatar is
loaded, and not as functions are called. Otherwise, any function
that was not called as part of the same page load as the new
avatar being uploaded will be missed.

Note - this repo now incorporates upstream changes to django-avatar.
Adding this to judgeapps requires a few other changes within apps,
to rename `django.avatar.util` to `utils`. Patch to follow.